midi notes track num uint64

+++ objects/midi_modernize/notes.py
@@ -8,7 +8,7 @@ import objects.midi_modernize.gfunc as gfunc
 
 midinote_premake = dynbytearr.dynbytearr_premake([
 	('complete', np.uint8),
-	('track', np.uint8),
+	('track', np.uint64),
 	('chanport', np.int32),
 	('start', np.uint64),
 	('end', np.uint64),
